# swagger.json
swagger: '2.0'
info:
  version: "1.0.1"
  title: Wardrobe
  description: Simple UI to promote .war files between S3 locations
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
basePath: "/v1"
...
schemes:
- "http"
paths:
...
  /shelf/{shelfId}/promote/{promoterId}:
    get:
      tags:
      - "shelf"
      summary: "Find shelf by ID"
      description: "Adds a new revision defined by the promoter"
      operationId: "promoteDrawerId"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "shelfId"
        in: "path"
        description: "ID of shelf to promote"
        required: true
        type: "integer"
        format: "int64"
      - name: "promoterId"
        in: "path"
        description: "ID of promoter to promote from"
        required: true
        type: "integer"
        format: "int64"
      responses:
        400:
          description: "Invalid ID supplied"
...
  /revision/:
    post: 
      tags:
      - "shelf"
      summary: "Create revision"
      description: "A version "
      operationId: "createRevision"
      consumes:
      - "application/xml"
      - "application/json"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Revision object that needs to be added to a shelf"
        required: true
        schema:
          $ref: "#/definitions/Revision"
      responses:
        405:
          description: "Invalid input"
  /revision/{revisionId}:
    get: 
      tags:
      - "shelf"
      parameters:
      - name: "revisionId"
        in: "path"
        description: "ID of shelf to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Revision"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Revision not found"
    put:
      tags:
      - "shelf"
      parameters:
      - name: "revisionId"
        in: "path"
        description: "ID of shelf to return"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        description: "Revision object that needs to be added to a shelf"
        required: true
        schema:
          $ref: "#/definitions/Revision"
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "Shelf not found"
        405:
          description: "Validation exception"
    delete:
      tags:
      - "shelf"
      parameters:
      - name: "revisionId"
        in: "path"
        description: "ID of shelf to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        405:
          description: "Invalid input"
...
  /promoter:
    post: 
      tags:
      - "shelf"
      summary: "Create promoter"
      description: "Creates a promoter object"
      operationId: "createPromoter"
      consumes:
      - "application/xml"
      - "application/json"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Drawer object that needs to be added to a shelf"
        required: true
        schema:
          $ref: "#/definitions/Promoter"
      responses:
        405:
          description: "Invalid input"
  /promoter/{promoterId}:
    get: 
      tags:
      - "shelf"
      parameters:
      - name: "promoterId"
        in: "path"
        description: "ID of promoter to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Promoter"
        400:
          description: "Invalid ID supplied"
        404:
          description: "promoter not found"
    put:
      tags:
      - "shelf"
      parameters:
      - name: "promoterId"
        in: "path"
        description: "ID of promoter to update"
        required: true
        type: "integer"
        format: "int64"
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "promoter not found"
        405:
          description: "Validation exception"
    delete:
      tags:
      - "shelf"
      parameters:
      - name: "promoterId"
        in: "path"
        description: "ID of promoter to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        405:
          description: "Invalid input"    
  /promote:
    post:
      tags:
      - "shelf"
      summary: "Promote a URI to a shelf"
      description: "Promotes a url to a shelf, creates a promoter \
      if url does not exist"
      operationId: "promoteShelf"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "wardrobeUrl"
        in: "query"
        description: "location of shelf to promote to"
        required: true
        type: "string"
      - name: "promoteUrl"
        in: "query"
        description: "uri to promote"
        required: true
        type: "string" 
      responses:
        405:
          description: "Invalid input"
  /rollback:
    post:
      tags:
      - "shelf"
      summary: "rollback a single item"
      description: "rollback one version previous"
      operationId: "rollbackShelf"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "wardrobeUrl"
        in: "query"
        description: "location of shelf to rollback"
        required: true
        type: "string"
      responses:
        405:
          description: "Invalid input"    
...
definitions:
...
  Revision:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
      dateAdded:
        type: "string"
        format: "date"
    xml:
      name: "Revision"
...
  Promoter:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      address:
        type: "string"
    xml:
      name: "Promoter"
...